requirements.txt
numpy
Pillow
ipywidgets


